Splitt opp barnehagelistekontrakt inn i separate filer
