+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/domene/Barnehage.kt
@@ -1,0 +1,8 @@
+package no.nav.familie.ks.barnehagelister.domene
+
+data class Barnehage(
+    val navn: String,
+    val organisasjonsnummer: String,
+    val adresse: Adresse?,
+    val barnInfolinjer: List<BarnInfolinje>,
+)
