+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/domene/Listeopplysninger.kt
@@ -1,0 +1,9 @@
+package no.nav.familie.ks.barnehagelister.domene
+
+import java.time.YearMonth
+
+data class Listeopplysninger(
+    val kommunenavn: String,
+    val kommunenummer: String,
+    val innsendingGjelderArManed: YearMonth,
+)
