+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/domene/Person.kt
@@ -1,0 +1,8 @@
+package no.nav.familie.ks.barnehagelister.domene
+
+data class Person(
+    val fodselsnummer: String,
+    val fornavn: String,
+    val etternavn: String,
+    val adresse: Adresse?,
+)
