+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/rest/DefaultBarnehagelisterController.kt
@@ -2,10 +2,10 @@ package no.nav.familie.ks.barnehagelister.rest
 
 import jakarta.servlet.http.HttpServletRequest
 import no.nav.familie.ks.barnehagelister.domene.BarnehagelisteService
-import no.nav.familie.ks.barnehagelister.domene.mapTilSkjemaV1
 import no.nav.familie.ks.barnehagelister.interceptor.hentSupplierId
 import no.nav.familie.ks.barnehagelister.rest.dto.FormV1RequestDto
 import no.nav.familie.ks.barnehagelister.rest.dto.KindergartenlistResponse
+import no.nav.familie.ks.barnehagelister.rest.dto.mapTilSkjemaV1
 import org.springframework.context.annotation.Profile
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.BindingResult
