+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/rest/dto/KindergartenRequestDto.kt
@@ -5,6 +5,9 @@ import jakarta.validation.Valid
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Size
+import no.nav.familie.ks.barnehagelister.domene.Barnehage
+import no.nav.familie.ks.barnehagelister.domene.mapTilAdresse
+import no.nav.familie.ks.barnehagelister.domene.mapTilBarnInfolinje
 
 data class KindergartenRequestDto(
     @Schema(description = "Name of the kindergarten")
@@ -25,3 +28,11 @@ data class KindergartenRequestDto(
     @field:Valid
     val childrenInformation: List<ChildInformationRequestDto>,
 )
+
+fun KindergartenRequestDto.mapTilBarnehage(): Barnehage =
+    Barnehage(
+        navn = this.name,
+        organisasjonsnummer = this.organizationNumber,
+        adresse = this.address?.mapTilAdresse(),
+        barnInfolinjer = this.childrenInformation.map { it.mapTilBarnInfolinje() },
+    )
